+++ examples/linkproppred/tgn.py
@@ -1,7 +1,7 @@
 import argparse
+import copy
 import time
-from collections import defaultdict
-from typing import Dict, Tuple
+from typing import Callable, Dict, Tuple
 
 import numpy as np
 import torch
@@ -9,9 +9,14 @@ import torch.nn as nn
 import torch.nn.functional as F
 from tgb.linkproppred.dataset_pyg import PyGLinkPropPredDataset
 from tgb.linkproppred.evaluate import Evaluator
+from torch import Tensor
+from torch.nn import GRUCell, Linear, RNNCell
+from torch_geometric.nn import TransformerConv
+from torch_geometric.nn.inits import zeros
+from torch_geometric.utils import scatter
 from tqdm import tqdm
 
-from tgm import DGBatch, DGData, DGraph
+from tgm import DGData, DGraph
 from tgm.constants import PADDED_NODE_ID
 from tgm.hooks import (
     HookManager,
@@ -21,21 +26,18 @@ from tgm.hooks import (
     TGBNegativeEdgeSamplerHook,
 )
 from tgm.loader import DGDataLoader
-from tgm.nn import TemporalAttention, Time2Vec
 from tgm.util.seed import seed_everything
 
 parser = argparse.ArgumentParser(
     description='TGN LinkPropPred Example',
     formatter_class=argparse.ArgumentDefaultsHelpFormatter,
 )
-parser.add_argument('--seed', type=int, default=1337, help='random seed to use')
+parser.add_argument('--seed', type=int, default=1, help='random seed to use')
 parser.add_argument('--dataset', type=str, default='tgbl-wiki', help='Dataset name')
 parser.add_argument('--bsize', type=int, default=200, help='batch size')
 parser.add_argument('--device', type=str, default='cpu', help='torch device')
-parser.add_argument('--epochs', type=int, default=10, help='number of epochs')
+parser.add_argument('--epochs', type=int, default=30, help='number of epochs')
 parser.add_argument('--lr', type=str, default=0.0001, help='learning rate')
-parser.add_argument('--dropout', type=str, default=0.1, help='dropout rate')
-parser.add_argument('--n-heads', type=int, default=2, help='number of attention heads')
 parser.add_argument(
     '--n-nbrs',
     type=int,
@@ -44,7 +46,8 @@ parser.add_argument(
     help='num sampled nbrs at each hop',
 )
 parser.add_argument('--time-dim', type=int, default=100, help='time encoding dimension')
-parser.add_argument('--embed-dim', type=int, default=172, help='attention dimension')
+parser.add_argument('--embed-dim', type=int, default=100, help='attention dimension')
+parser.add_argument('--memory-dim', type=int, default=100, help='memory dimension')
 parser.add_argument(
     '--sampling',
     type=str,
@@ -52,358 +55,374 @@ parser.add_argument(
     choices=['uniform', 'recency'],
     help='sampling strategy',
 )
-parser.add_argument(
-    '--node-dim', type=int, default=256, help='node feat dimension if not provided'
-)
 
 
-class MergeLayer(nn.Module):
-    def __init__(self, in_dim1: int, in_dim2: int, hidden_dim: int, output_dim: int):
+# -------------------------------------------------------
+class LinkPredictor(torch.nn.Module):
+    def __init__(self, in_channels):
         super().__init__()
-        self.fc1 = nn.Linear(in_dim1 + in_dim2, hidden_dim)
-        self.fc2 = nn.Linear(hidden_dim, output_dim)
+        self.lin_src = Linear(in_channels, in_channels)
+        self.lin_dst = Linear(in_channels, in_channels)
+        self.lin_final = Linear(in_channels, 1)
 
-    def forward(self, x1: torch.Tensor, x2: torch.Tensor):
-        h = self.fc1(torch.cat([x1, x2], dim=1))
+    def forward(self, z_src, z_dst):
+        h = self.lin_src(z_src) + self.lin_dst(z_dst)
         h = h.relu()
-        return self.fc2(h)
+        return self.lin_final(h).sigmoid()
 
 
-class TGN(torch.nn.Module):
+class GraphAttentionEmbedding(torch.nn.Module):
+    def __init__(self, in_channels, out_channels, msg_dim, time_enc):
+        super().__init__()
+        self.time_enc = time_enc
+        edge_dim = msg_dim + time_enc.out_channels
+        self.conv = TransformerConv(
+            in_channels, out_channels // 2, heads=2, dropout=0.1, edge_dim=edge_dim
+        )
+
+    def forward(self, x, last_update, edge_index, t, msg):
+        rel_t = last_update[edge_index[0]] - t
+        rel_t_enc = self.time_enc(rel_t.to(x.dtype))
+        edge_attr = torch.cat([rel_t_enc, msg], dim=-1)
+        return self.conv(x, edge_index, edge_attr)
+
+
+class LastAggregator(torch.nn.Module):
+    def forward(self, msg: Tensor, index: Tensor, t: Tensor, dim_size: int):
+        out = msg.new_zeros((dim_size, msg.size(-1)))
+
+        # _, argmax = scatter_max(t, index, dim=0, dim_size=dim_size)
+        # mask = argmax < msg.size(0)  # Filter items with at least one entry.
+        # out[mask] = msg[argmax[mask]]
+
+        for i in range(dim_size):
+            mask = index == i
+            if mask.any():
+                local_idx = torch.argmax(t[mask])
+                global_idx = mask.nonzero(as_tuple=True)[0][local_idx]
+                out[i] = msg[global_idx]
+
+        return out
+
+
+class IdentityMessage(torch.nn.Module):
+    def __init__(self, raw_msg_dim: int, memory_dim: int, time_dim: int):
+        super().__init__()
+        self.out_channels = raw_msg_dim + 2 * memory_dim + time_dim
+
+    def forward(self, z_src: Tensor, z_dst: Tensor, raw_msg: Tensor, t_enc: Tensor):
+        return torch.cat([z_src, z_dst, raw_msg, t_enc], dim=-1)
+
+
+TGNMessageStoreType = Dict[int, Tuple[Tensor, Tensor, Tensor, Tensor]]
+
+
+class TGNMemory(torch.nn.Module):
     def __init__(
         self,
-        node_dim: int,
-        edge_dim: int,
-        time_dim: int,
-        embed_dim: int,
-        num_layers: int,
-        n_heads: int,
-        dropout: float,
         num_nodes: int,
+        raw_msg_dim: int,
+        memory_dim: int,
+        time_dim: int,
+        message_module: Callable,
+        aggregator_module: Callable,
+        memory_updater_cell: str = 'gru',
     ):
         super().__init__()
-        self.time_encoder = Time2Vec(time_dim=time_dim)
-        self.msg_agg = LastMessageAggregator()
-        self.nodes = list(range(num_nodes))
 
-        msg_dim = 2 * embed_dim + edge_dim + time_dim
-        self.memory = Memory(n_nodes=num_nodes, memory_dim=embed_dim)
-        self.memory_updater = GRUMemoryUpdater(self.memory, msg_dim, embed_dim)
-        self.gat = GraphAttentionEmbedding(
-            node_dim=node_dim,
-            edge_dim=edge_dim,
-            time_dim=time_dim,
-            embed_dim=embed_dim,
-            num_layers=num_layers,
-            n_heads=n_heads,
-            dropout=dropout,
-        )
-
-    def forward(
-        self, batch: DGBatch, static_node_feats: torch.Tensor
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        agg_msgs = self.msg_agg(self.nodes, self.memory.msgs)
-        memory, _ = self.memory_updater.get_updated_memory(*agg_msgs)
-        # TODO: I think this is only needed for multi-hop?
-        # Difference between the time the memory of a node was last updated,
-        # and the time for which we want to compute the embedding of a node
-        # batch.time[batch.src] -= last_update[batch.src].long()
-        # batch.time[batch.dst] -= last_update[batch.dst].long()
-        # batch.time[batch.neg] -= last_update[batch.neg].long()
-
-        z = self.gat(batch, static_node_feats, memory=memory)
-        self._update_memory(batch)
-        return z
-
-    def _update_memory(self, batch: DGBatch) -> None:
-        device = batch.src.device
-
-        def _get_raw_msgs(src, dst, time):
-            edge_feats = batch.edge_feats
-            src_memory = self.memory.get_memory(src)
-            dst_memory = self.memory.get_memory(dst)
-            time_delta = time.to(device) - self.memory.last_update[src]
-            time_feat = self.time_encoder(time_delta.unsqueeze(dim=1)).view(
-                len(src), -1
-            )
-
-            src_msg = torch.cat([src_memory, dst_memory, edge_feats, time_feat], dim=1)
-            msgs = defaultdict(list)
-            unique_src = np.unique(src.cpu())
-            for i in range(len(src)):
-                msgs[src[i]].append((src_msg[i], time[i]))
-            return unique_src, msgs
-
-        # Persist the updates to the memory only for sources and destinations
-        pos = torch.cat([batch.src, batch.dst]).cpu().numpy()
-        agg_msgs = self.msg_agg(pos, self.memory.msgs)
-        self.memory_updater.update(*agg_msgs)
-
-        # Remove msgs for the pos since we have already updated the memory using them
-        self.memory.clear_msgs(pos)
-        unique_src, src_to_msgs = _get_raw_msgs(batch.src, batch.dst, batch.time)
-        unique_dst, dst_to_msgs = _get_raw_msgs(batch.dst, batch.src, batch.time)
-        self.memory.store_raw_msgs(unique_src, src_to_msgs)
-        self.memory.store_raw_msgs(unique_dst, dst_to_msgs)
-
-
-class LastMessageAggregator(torch.nn.Module):
-    def forward(
-        self, node_ids: torch.Tensor, msgs: torch.Tensor
-    ) -> Tuple[torch.Tensor, ...]:
-        unique_nids = np.unique(node_ids)
-        unique_msg, unique_times, to_update_node_ids = [], [], []
-        for node_id in unique_nids:
-            if len(msgs[node_id]) > 0:
-                to_update_node_ids.append(node_id)
-                unique_msg.append(msgs[node_id][-1][0])
-                unique_times.append(msgs[node_id][-1][1])
-        unique_msg = torch.stack(unique_msg) if len(to_update_node_ids) > 0 else []
-        unique_times = torch.stack(unique_times) if len(to_update_node_ids) > 0 else []
-        return to_update_node_ids, unique_msg, unique_times
-
-
-class Memory(nn.Module):
-    def __init__(self, n_nodes: int, memory_dim: int, device: str = 'cpu') -> None:
-        super().__init__()
+        self.num_nodes = num_nodes
+        self.raw_msg_dim = raw_msg_dim
         self.memory_dim = memory_dim
-        self.n_nodes = n_nodes
-        self.device = device
-        self.msgs = defaultdict(list)
+        self.time_dim = time_dim
 
-        self.reset()
+        self.msg_s_module = message_module
+        self.msg_d_module = copy.deepcopy(message_module)
+        self.aggr_module = aggregator_module
+        self.time_enc = TimeEncoder(time_dim)
+        if memory_updater_cell == 'gru':  # for TGN
+            self.memory_updater = GRUCell(message_module.out_channels, memory_dim)
+        elif memory_updater_cell == 'rnn':  # for JODIE & DyRep
+            self.memory_updater = RNNCell(message_module.out_channels, memory_dim)
+        else:
+            raise ValueError(
+                "Undefined memory updater!!! Memory updater can be either 'gru' or 'rnn'."
+            )
 
-    def set_device(self, device: str) -> None:
-        self.device = device
+        self.register_buffer('memory', torch.empty(num_nodes, memory_dim))
+        last_update = torch.empty(self.num_nodes, dtype=torch.long)
+        self.register_buffer('last_update', last_update)
+        self.register_buffer('_assoc', torch.empty(num_nodes, dtype=torch.long))
 
-    def reset(self) -> None:
-        self.memory = nn.Parameter(
-            torch.zeros((self.n_nodes, self.memory_dim), device=self.device),
-            requires_grad=False,
-        )
-        self.last_update = nn.Parameter(
-            torch.zeros(self.n_nodes, device=self.device), requires_grad=False
-        )
+        self.msg_s_store = {}
+        self.msg_d_store = {}
 
-    def store_raw_msgs(
-        self, nodes: torch.Tensor, node_id_to_msg: Dict[int, torch.Tensor]
-    ) -> None:
-        for node in nodes:
-            self.msgs[node].extend(node_id_to_msg[node])
+        self.reset_parameters()
 
-    def get_memory(self, node_idxs: torch.Tensor) -> torch.Tensor:
-        return self.memory[node_idxs, :]
+    @property
+    def device(self) -> torch.device:
+        return self.time_enc.lin.weight.device
 
-    def set_memory(self, node_idxs: torch.Tensor, values: torch.Tensor) -> None:
-        self.memory[node_idxs, :] = values
+    def reset_parameters(self):
+        if hasattr(self.msg_s_module, 'reset_parameters'):
+            self.msg_s_module.reset_parameters()
+        if hasattr(self.msg_d_module, 'reset_parameters'):
+            self.msg_d_module.reset_parameters()
+        if hasattr(self.aggr_module, 'reset_parameters'):
+            self.aggr_module.reset_parameters()
+        self.time_enc.reset_parameters()
+        self.memory_updater.reset_parameters()
+        self.reset_state()
 
-    def get_last_update(self, node_idxs: torch.Tensor) -> torch.Tensor:
-        return self.last_update[node_idxs]
+    def reset_state(self):
+        zeros(self.memory)
+        zeros(self.last_update)
+        self._reset_message_store()
 
-    def backup_memory(self) -> Tuple[torch.Tensor, ...]:
-        msgs_clone = {}
-        for k, v in self.msgs.items():
-            msgs_clone[k] = [(x[0].clone(), x[1].clone()) for x in v]
-        return self.memory.data.clone(), self.last_update.data.clone(), msgs_clone
-
-    def restore_memory(self, memory_backup: Tuple[torch.Tensor, ...]) -> None:
-        self.memory.data, self.last_update.data = (
-            memory_backup[0].clone(),
-            memory_backup[1].clone(),
-        )
-        self.msgs = defaultdict(list)
-        for k, v in memory_backup[2].items():
-            self.msgs[k] = [(x[0].clone(), x[1].clone()) for x in v]
-
-    def detach_memory(self) -> None:
+    def detach(self):
         self.memory.detach_()
-        for k, v in self.msgs.items():
-            new_node_msg = []
-            for msg in v:
-                new_node_msg.append((msg[0].detach(), msg[1]))
-            self.msgs[k] = new_node_msg
 
-    def clear_msgs(self, nodes: torch.Tensor) -> None:
-        for node in nodes:
-            self.msgs[node] = []
+    def forward(self, n_id: Tensor) -> Tuple[Tensor, Tensor]:
+        if self.training:
+            memory, last_update = self._get_updated_memory(n_id)
+        else:
+            memory, last_update = self.memory[n_id], self.last_update[n_id]
 
+        return memory, last_update
 
-class GRUMemoryUpdater(nn.Module):
-    def __init__(self, memory: Memory, msg_dim: int, memory_dim: int) -> None:
-        super().__init__()
-        self.memory = memory
-        self.memory_updater = nn.GRUCell(input_size=msg_dim, hidden_size=memory_dim)
+    def update_state(self, src: Tensor, dst: Tensor, t: Tensor, raw_msg: Tensor):
+        n_id = torch.cat([src, dst]).unique()
 
-    def update(
-        self, unique_nids: torch.Tensor, unique_msg: torch.Tensor, time: int
-    ) -> None:
-        if len(unique_nids) <= 0:
-            return
-        memory = self.memory.get_memory(unique_nids)
-        self.memory.last_update[unique_nids] = time
-        updated_memory = self.memory_updater(unique_msg, memory)
-        self.memory.set_memory(unique_nids, updated_memory)
+        if self.training:
+            self._update_memory(n_id)
+            self._update_msg_store(src, dst, t, raw_msg, self.msg_s_store)
+            self._update_msg_store(dst, src, t, raw_msg, self.msg_d_store)
+        else:
+            self._update_msg_store(src, dst, t, raw_msg, self.msg_s_store)
+            self._update_msg_store(dst, src, t, raw_msg, self.msg_d_store)
+            self._update_memory(n_id)
 
-    def get_updated_memory(
-        self, unique_nids: torch.Tensor, unique_msg: torch.Tensor, time: int
-    ) -> Tuple[torch.Tensor, torch.Tensor]:
-        if len(unique_nids) <= 0:
-            return self.memory.memory.data.clone(), self.memory.last_update.data.clone()
-        updated_memory = self.memory.memory.data.clone()
-        updated_memory[unique_nids] = self.memory_updater(
-            unique_msg, updated_memory[unique_nids]
+    def _reset_message_store(self):
+        i = self.memory.new_empty((0,), device=self.device, dtype=torch.long)
+        msg = self.memory.new_empty((0, self.raw_msg_dim), device=self.device)
+        # Message store format: (src, dst, t, msg)
+        self.msg_s_store = {j: (i, i, i, msg) for j in range(self.num_nodes)}
+        self.msg_d_store = {j: (i, i, i, msg) for j in range(self.num_nodes)}
+
+    def _update_memory(self, n_id: Tensor):
+        memory, last_update = self._get_updated_memory(n_id)
+        self.memory[n_id] = memory
+        self.last_update[n_id] = last_update
+
+    def _get_updated_memory(self, n_id: Tensor) -> Tuple[Tensor, Tensor]:
+        self._assoc[n_id] = torch.arange(n_id.size(0), device=n_id.device)
+
+        # Compute messages (src -> dst).
+        msg_s, t_s, src_s, dst_s = self._compute_msg(
+            n_id, self.msg_s_store, self.msg_s_module
         )
-        updated_last_update = self.memory.last_update.data.clone()
-        updated_last_update[unique_nids] = time
-        return updated_memory, updated_last_update
 
+        # Compute messages (dst -> src).
+        msg_d, t_d, src_d, dst_d = self._compute_msg(
+            n_id, self.msg_d_store, self.msg_d_module
+        )
 
-class GraphAttentionEmbedding(nn.Module):
-    def __init__(
+        # Aggregate messages.
+        idx = torch.cat([src_s, src_d], dim=0)
+        msg = torch.cat([msg_s, msg_d], dim=0)
+        t = torch.cat([t_s, t_d], dim=0)
+        aggr = self.aggr_module(msg, self._assoc[idx], t, n_id.size(0))
+
+        # Get local copy of updated memory.
+        memory = self.memory_updater(aggr, self.memory[n_id])
+
+        # Get local copy of updated `last_update`.
+        dim_size = self.last_update.size(0)
+        last_update = scatter(t, idx, 0, dim_size, reduce='max')[n_id]
+        return memory, last_update
+
+    def _update_msg_store(
         self,
-        node_dim: int,
-        edge_dim: int,
-        time_dim: int,
-        embed_dim: int,
-        num_layers: int,
-        n_heads: int = 2,
-        dropout: float = 0.1,
-    ) -> None:
-        """In this implementation, the node embedding dimension must be the same as hidden embedding dimension."""
+        src: Tensor,
+        dst: Tensor,
+        t: Tensor,
+        raw_msg: Tensor,
+        msg_store: TGNMessageStoreType,
+    ):
+        n_id, perm = src.sort()
+        n_id, count = n_id.unique_consecutive(return_counts=True)
+        for i, idx in zip(n_id.tolist(), perm.split(count.tolist())):
+            msg_store[i] = (src[idx], dst[idx], t[idx], raw_msg[idx])
+
+    def _compute_msg(
+        self, n_id: Tensor, msg_store: TGNMessageStoreType, msg_module: Callable
+    ):
+        data = [msg_store[i] for i in n_id.tolist()]
+        src, dst, t, raw_msg = list(zip(*data))
+        src = torch.cat(src, dim=0)
+        dst = torch.cat(dst, dim=0)
+        t = torch.cat(t, dim=0)
+        raw_msg = torch.cat(raw_msg, dim=0)
+        t_rel = t - self.last_update[src]
+        t_enc = self.time_enc(t_rel.to(raw_msg.dtype))
+        msg = msg_module(self.memory[src], self.memory[dst], raw_msg, t_enc)
+        return msg, t, src, dst
+
+    def train(self, mode: bool = True):
+        if self.training and not mode:
+            # Flush message store to memory in case we just entered eval mode.
+            self._update_memory(torch.arange(self.num_nodes, device=self.memory.device))
+            self._reset_message_store()
+        super().train(mode)
+
+
+class TimeEncoder(torch.nn.Module):
+    def __init__(self, out_channels: int):
         super().__init__()
-        self.num_layers = num_layers
-        self.embed_dim = embed_dim
-        self.time_encoder = Time2Vec(time_dim=time_dim)
+        self.out_channels = out_channels
+        self.lin = Linear(1, out_channels)
 
-        self.attn, self.merge_layers = nn.ModuleList(), nn.ModuleList()
-        for i in range(num_layers):
-            self.attn.append(
-                TemporalAttention(
-                    n_heads=n_heads,
-                    node_dim=node_dim if i == 0 else embed_dim,
-                    edge_dim=edge_dim,
-                    time_dim=time_dim,
-                    dropout=dropout,
-                )
-            )
-            self.merge_layers.append(
-                MergeLayer(
-                    in_dim1=self.attn[-1].out_dim,
-                    in_dim2=node_dim,
-                    hidden_dim=embed_dim,
-                    output_dim=embed_dim,
-                )
-            )
+    def reset_parameters(self):
+        self.lin.reset_parameters()
 
-    def forward(
-        self, batch: DGBatch, static_node_feat: torch.Tensor, memory: Memory
-    ) -> torch.Tensor:
-        device = batch.src.device
-        z = {j: {} for j in range(self.num_layers + 1)}  # z[j][i] = z of nbr^i at hop j
-
-        # Layer 0 (leaf nodes): z[0][i] = static_node_feat
-        z[0][0] = static_node_feat[batch.nids[0]]
-        z[0][0] += memory[batch.nids[0]]
-        for i in range(1, self.num_layers + 1):
-            z[0][i] = static_node_feat[batch.nbr_nids[i - 1].flatten()]
-            z[0][i] += memory[batch.nbr_nids[i - 1].flatten()]
-
-        # Layers 1..H: aggregate z[j][i] = agg(z[j - 1][i], z[j - 1][i + 1])
-        for j in range(1, self.num_layers + 1):
-            for i in range(self.num_layers - j + 1):
-                num_nodes = z[j - 1][i].size(0)
-                num_nbr = batch.nbr_nids[j - 1].shape[-1]
-                out = self.attn[j - 1](
-                    node_feat=z[j - 1][i],
-                    time_feat=self.time_encoder(torch.zeros(num_nodes, device=device)),
-                    nbr_node_feat=z[j - 1][i + 1].reshape(num_nodes, num_nbr, -1),
-                    edge_feat=batch.nbr_feats[i],
-                    valid_nbr_mask=batch.nbr_nids[i] != PADDED_NODE_ID,
-                    nbr_time_feat=self.time_encoder(
-                        batch.times[i][:, None] - batch.nbr_times[i]
-                    ),
-                )
-                z[j][i] = self.merge_layers[j - 1](out, z[0][i])
-
-        return z[self.num_layers][0]
+    def forward(self, t: Tensor) -> Tensor:
+        return self.lin(t.view(-1, 1)).cos()
 
 
-class LinkPredictor(nn.Module):
-    def __init__(self, dim: int) -> None:
-        super().__init__()
-        self.fc1 = nn.Linear(2 * dim, dim)
-        self.fc2 = nn.Linear(dim, 1)
-
-    def forward(self, z_src: torch.Tensor, z_dst: torch.Tensor) -> torch.Tensor:
-        h = self.fc1(torch.cat([z_src, z_dst], dim=1))
-        h = h.relu()
-        return self.fc2(h).sigmoid().view(-1)
+# -------------------------------------------------------
 
 
 def train(
     loader: DGDataLoader,
     static_node_feats: torch.Tensor,
+    memory: nn.Module,
     encoder: nn.Module,
     decoder: nn.Module,
     opt: torch.optim.Optimizer,
 ) -> float:
+    memory.train()
     encoder.train()
     decoder.train()
     total_loss = 0
 
-    encoder.memory.reset()  # Re-init memory of the model at the start of each epoch
-
     for batch in tqdm(loader):
         opt.zero_grad()
 
-        z = encoder(batch, static_node_feats)
-        z_src, z_dst, z_neg = torch.chunk(z, 3)
+        nbr_nodes = batch.nbr_nids[0].flatten()
+        nbr_mask = nbr_nodes != PADDED_NODE_ID  # mask out invalid nbrs
 
-        pos_out = decoder(z_src, z_dst)
-        neg_out = decoder(z_src, z_neg)
+        #! run my own deduplication
+        nids = [batch.src, batch.dst, batch.neg, nbr_nodes[nbr_mask]]
+        all_nids = torch.cat(nids, dim=0)
+        unique_nids = torch.unique(all_nids, sorted=True)
+        batch.unique_nids = unique_nids  # type: ignore
+        batch.global_to_local = lambda x: torch.searchsorted(unique_nids, x)  # type: ignore
+
+        num_nbrs = len(nbr_nodes) // (len(batch.src) + len(batch.dst) + len(batch.neg))
+        src_nodes = torch.cat(
+            (
+                batch.src.repeat_interleave(num_nbrs),
+                batch.dst.repeat_interleave(num_nbrs),
+                batch.neg.repeat_interleave(num_nbrs),
+            ),
+            0,
+        )
+        nbr_edge_index = torch.stack(
+            [
+                batch.global_to_local(src_nodes[nbr_mask]),
+                batch.global_to_local(nbr_nodes[nbr_mask]),
+            ]
+        )
+
+        nbr_times = batch.nbr_times[0].flatten()[nbr_mask]
+        nbr_feats = batch.nbr_feats[0].flatten(0, -2).float()[nbr_mask]
+        z, last_update = memory(batch.unique_nids)
+        z = encoder(z, last_update, nbr_edge_index, nbr_times, nbr_feats)
+        pos_out = decoder(
+            z[batch.global_to_local(batch.src)], z[batch.global_to_local(batch.dst)]
+        )
+        neg_out = decoder(
+            z[batch.global_to_local(batch.src)], z[batch.global_to_local(batch.neg)]
+        )
 
         loss = F.binary_cross_entropy_with_logits(pos_out, torch.ones_like(pos_out))
         loss += F.binary_cross_entropy_with_logits(neg_out, torch.zeros_like(neg_out))
+
+        # Update memory and neighbor loader with ground-truth state.
+        memory.update_state(batch.src, batch.dst, batch.time, batch.edge_feats.float())
+
         loss.backward()
         opt.step()
         total_loss += float(loss)
 
-        encoder.memory.detach_memory()  # Detach memory to avoid BPTT
+        memory.detach()
+
     return total_loss
 
 
 @torch.no_grad()
 def eval(
-    loader: DGDataLoader,
+    loader,
     static_node_feats: torch.Tensor,
+    memory: nn.Module,
     encoder: nn.Module,
     decoder: nn.Module,
     eval_metric: str,
     evaluator: Evaluator,
 ) -> float:
+    memory.eval()
     encoder.eval()
     decoder.eval()
     perf_list = []
 
     for batch in tqdm(loader):
-        z = encoder(batch, static_node_feats)
-        id_map = {nid.item(): i for i, nid in enumerate(batch.nids[0])}
+        nbr_nodes = batch.nbr_nids[0].flatten()
+        nbr_mask = nbr_nodes != PADDED_NODE_ID  # mask out invalid nbrs
+
+        #! run my own deduplication
+        nids = [batch.src, batch.dst, batch.neg, nbr_nodes[nbr_mask]]
+        all_nids = torch.cat(nids, dim=0)
+        unique_nids = torch.unique(all_nids, sorted=True)
+        batch.unique_nids = unique_nids  # type: ignore
+        batch.global_to_local = lambda x: torch.searchsorted(unique_nids, x)  # type: ignore
+
+        nbr_nodes = batch.nbr_nids[0].flatten()
+        num_nbrs = len(nbr_nodes) // (len(batch.src) + len(batch.dst) + len(batch.neg))
+        src_nodes = torch.cat(
+            (
+                batch.src.repeat_interleave(num_nbrs),
+                batch.dst.repeat_interleave(num_nbrs),
+                batch.neg.repeat_interleave(num_nbrs),
+            ),
+            0,
+        )
+        nbr_mask = nbr_nodes != -1
+        nbr_edge_index = torch.stack(
+            [
+                batch.global_to_local(src_nodes[nbr_mask]),
+                batch.global_to_local(nbr_nodes[nbr_mask]),
+            ]
+        )
+        nbr_times = batch.nbr_times[0].flatten()[nbr_mask]
+        nbr_feats = batch.nbr_feats[0].flatten(0, -2).float()[nbr_mask]
+
+        z, last_update = memory(batch.unique_nids)
+        z = encoder(z, last_update, nbr_edge_index, nbr_times, nbr_feats)
+
         for idx, neg_batch in enumerate(batch.neg_batch_list):
-            dst_ids = torch.cat([batch.dst[idx].unsqueeze(0), neg_batch])
-            src_ids = batch.src[idx].repeat(len(dst_ids))
-
-            src_idx = torch.tensor([id_map[n.item()] for n in src_ids], device=z.device)
-            dst_idx = torch.tensor([id_map[n.item()] for n in dst_ids], device=z.device)
-            z_src = z[src_idx]
-            z_dst = z[dst_idx]
-            y_pred = decoder(z_src, z_dst)
-
+            dst = torch.cat([batch.dst[idx].unsqueeze(0), neg_batch])
+            src = batch.src[idx].repeat(len(dst))
+            y_pred = decoder(
+                z[batch.global_to_local(src)], z[batch.global_to_local(dst)]
+            )
             input_dict = {
-                'y_pred_pos': y_pred[0],
-                'y_pred_neg': y_pred[1:],
+                'y_pred_pos': np.array([y_pred[0, :].squeeze(dim=-1).cpu()]),
+                'y_pred_neg': np.array(y_pred[1:, :].squeeze(dim=-1).cpu()),
                 'eval_metric': [eval_metric],
             }
             perf_list.append(evaluator.eval(input_dict)[eval_metric])
+
+        # Update memory with ground-truth state.
+        memory.update_state(batch.src, batch.dst, batch.time, batch.edge_feats.float())
 
     return float(np.mean(perf_list))
 
@@ -426,9 +445,7 @@ test_dg = DGraph(test_data, device=args.device)
 if train_dg.static_node_feats is not None:
     static_node_feats = train_dg.static_node_feats
 else:
-    static_node_feats = torch.zeros(
-        (test_dg.num_nodes, args.node_dim), device=args.device
-    )
+    static_node_feats = torch.zeros((test_dg.num_nodes, 1), device=args.device)
 
 if args.sampling == 'uniform':
     nbr_hook = NeighborSamplerHook(num_nbrs=args.n_nbrs)
@@ -441,9 +458,7 @@ elif args.sampling == 'recency':
 else:
     raise ValueError(f'Unknown sampling type: {args.sampling}')
 
-
 _, dst, _ = train_dg.edges
-
 hm = HookManager(keys=['train', 'val', 'test'])
 hm.register('train', NegativeEdgeSamplerHook(low=int(dst.min()), high=int(dst.max())))
 hm.register('val', TGBNegativeEdgeSamplerHook(neg_sampler, split_mode='val'))
@@ -454,32 +469,45 @@ train_loader = DGDataLoader(train_dg, args.bsize, hook_manager=hm)
 val_loader = DGDataLoader(val_dg, args.bsize, hook_manager=hm)
 test_loader = DGDataLoader(test_dg, args.bsize, hook_manager=hm)
 
-encoder = TGN(
-    node_dim=static_node_feats.shape[1],
-    edge_dim=train_dg.edge_feats_dim,
-    time_dim=args.time_dim,
-    embed_dim=args.embed_dim,
-    num_layers=len(args.n_nbrs),
-    n_heads=args.n_heads,
-    dropout=float(args.dropout),
-    num_nodes=test_dg.num_nodes,
+
+memory = TGNMemory(
+    test_dg.num_nodes,
+    test_dg.edge_feats_dim,
+    args.memory_dim,
+    args.time_dim,
+    message_module=IdentityMessage(
+        test_dg.edge_feats_dim, args.memory_dim, args.time_dim
+    ),
+    aggregator_module=LastAggregator(),
 ).to(args.device)
-encoder.memory.set_device(args.device)
-decoder = LinkPredictor(dim=args.embed_dim).to(args.device)
+encoder = GraphAttentionEmbedding(
+    in_channels=args.memory_dim,
+    out_channels=args.embed_dim,
+    msg_dim=test_dg.edge_feats_dim,
+    time_enc=memory.time_enc,
+).to(args.device)
+decoder = LinkPredictor(in_channels=args.embed_dim).to(args.device)
 opt = torch.optim.Adam(
-    set(encoder.parameters()) | set(decoder.parameters()), lr=float(args.lr)
+    set(memory.parameters()) | set(encoder.parameters()) | set(decoder.parameters()),
+    lr=args.lr,
 )
 
 for epoch in range(1, args.epochs + 1):
     with hm.activate('train'):
         start_time = time.perf_counter()
-        loss = train(train_loader, static_node_feats, encoder, decoder, opt)
+        loss = train(train_loader, static_node_feats, memory, encoder, decoder, opt)
         end_time = time.perf_counter()
         latency = end_time - start_time
 
     with hm.activate('val'):
         val_mrr = eval(
-            val_loader, static_node_feats, encoder, decoder, eval_metric, evaluator
+            val_loader,
+            static_node_feats,
+            memory,
+            encoder,
+            decoder,
+            eval_metric,
+            evaluator,
         )
     print(
         f'Epoch={epoch:02d} Latency={latency:.4f} Loss={loss:.4f} Validation {eval_metric}={val_mrr:.4f}'
@@ -488,11 +516,11 @@ for epoch in range(1, args.epochs + 1):
     # Clear memory state between epochs, except last epoch
     if epoch < args.epochs:
         hm.reset_state()
-        encoder.memory.clear_msgs(list(range(test_dg.num_nodes)))
+        memory.reset_state()
 
 
 with hm.activate('test'):
     test_mrr = eval(
-        test_loader, static_node_feats, encoder, decoder, eval_metric, evaluator
+        test_loader, static_node_feats, memory, encoder, decoder, eval_metric, evaluator
     )
     print(f'Test {eval_metric}={test_mrr:.4f}')
